+++ moonmining/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "moonmining.apps.MoonPlanerConfig"
 
-__version__ = "1.8.2"
+__version__ = "1.9.0"
 __title__ = "Moon Mining"