+++ moonmining/views.py
@@ -1,6 +1,6 @@
 import datetime as dt
-from collections import defaultdict
 from enum import Enum
+from typing import Union
 
 from django_datatables_view.base_datatable_view import BaseDatatableView
 
@@ -29,6 +29,7 @@ from django.shortcuts import redirect, render
 from django.urls import reverse
 from django.utils.html import format_html, strip_tags
 from django.utils.timezone import now
+from django.utils.translation import gettext_lazy as _
 from django.views.decorators.cache import cache_page
 from esi.decorators import token_required
 
@@ -72,9 +73,10 @@ def moon_link_html(moon: Moon) -> str:
     return format_html(
         '<a href="#" data-toggle="modal" '
         'data-target="#modalMoonDetails" '
-        'title="Show details for this moon." '
+        'title="{}" '
         "data-ajax_url={}>"
         "{}</a>",
+        _("Show details for this moon."),
         reverse("moonmining:moon_details", args=[moon.pk]),
         moon.name,
     )
@@ -94,7 +96,7 @@ def moon_details_button_html(moon: Moon) -> str:
         modal_id="modalMoonDetails",
         fa_code="fas fa-moon",
         ajax_url=reverse("moonmining:moon_details", args=[moon.pk]),
-        tooltip="Moon details",
+        tooltip=_("Moon details"),
     )
 
 
@@ -103,7 +105,7 @@ def extraction_details_button_html(extraction_pk: int) -> str:
         modal_id="modalExtractionDetails",
         fa_code="fas fa-hammer",
         ajax_url=reverse("moonmining:extraction_details", args=[extraction_pk]),
-        tooltip="Extraction details",
+        tooltip=_("Extraction details"),
     )
 
 
@@ -133,7 +135,7 @@ def index(request):
 @permission_required(["moonmining.extractions_access", "moonmining.basic_access"])
 def extractions(request):
     context = {
-        "page_title": "Extractions",
+        "page_title": _("Extractions"),
         "ExtractionsCategory": ExtractionsCategory.to_dict(),
         "ExtractionsStatus": Extraction.Status,
         "use_reprocess_pricing": MOONMINING_USE_REPROCESS_PRICING,
@@ -148,10 +150,10 @@ def extractions(request):
 @permission_required(["moonmining.extractions_access", "moonmining.basic_access"])
 def extractions_data(request, category):
     data = list()
-    cutover_dt = now() - dt.timedelta(
+    stale_cutoff = now() - dt.timedelta(
         hours=MOONMINING_COMPLETED_EXTRACTIONS_HOURS_UNTIL_STALE
     )
-    extractions = (
+    extractions_qs = (
         Extraction.objects.annotate_volume()
         .selected_related_defaults()
         .select_related(
@@ -161,17 +163,17 @@ def extractions_data(request, category):
         )
     )
     if category == ExtractionsCategory.UPCOMING:
-        extractions = extractions.filter(auto_fracture_at__gte=cutover_dt).exclude(
-            status=Extraction.Status.CANCELED
-        )
+        extractions_qs = extractions_qs.filter(
+            auto_fracture_at__gte=stale_cutoff
+        ).exclude(status=Extraction.Status.CANCELED)
     elif category == ExtractionsCategory.PAST:
-        extractions = extractions.filter(
-            auto_fracture_at__lt=cutover_dt
-        ) | extractions.filter(status=Extraction.Status.CANCELED)
+        extractions_qs = extractions_qs.filter(
+            auto_fracture_at__lt=stale_cutoff
+        ) | extractions_qs.filter(status=Extraction.Status.CANCELED)
     else:
-        extractions = Extraction.objects.none()
+        extractions_qs = Extraction.objects.none()
     can_see_ledger = request.user.has_perm("moonmining.view_moon_ledgers")
-    for extraction in extractions:
+    for extraction in extractions_qs:
         corporation_name = extraction.refinery.owner.name
         alliance_name = extraction.refinery.owner.alliance_name
         moon = extraction.refinery.moon
@@ -270,7 +272,7 @@ def extraction_details(request, extraction_pk: int):
         "extraction": extraction,
     }
     if request.GET.get("new_page"):
-        context["title"] = "Extraction"
+        context["title"] = _("Extraction")
         context["content_file"] = "moonmining/partials/extraction_details.html"
         return render(request, "moonmining/_generic_modal_page.html", context)
     else:
@@ -342,7 +344,7 @@ def extraction_ledger(request, extraction_pk: int):
         "character_totals": character_totals,
     }
     if request.GET.get("new_page"):
-        context["title"] = "Extraction Ledger"
+        context["title"] = _("Extraction Ledger")
         context["content_file"] = "moonmining/partials/extraction_ledger.html"
         return render(request, "moonmining/_generic_modal_page.html", context)
     return render(request, "moonmining/modals/extraction_ledger.html", context)
@@ -352,7 +354,7 @@ def extraction_ledger(request, extraction_pk: int):
 @permission_required("moonmining.basic_access")
 def moons(request):
     context = {
-        "page_title": "Moons",
+        "page_title": _("Moons"),
         "MoonsCategory": MoonsCategory.to_dict(),
         "use_reprocess_pricing": MOONMINING_USE_REPROCESS_PRICING,
         "reprocessing_yield": MOONMINING_REPROCESSING_YIELD * 100,
@@ -503,21 +505,21 @@ class MoonListJson(PermissionRequiredMixin, LoginRequiredMixin, BaseDatatableVie
             )
         return qs
 
-        qs = self._apply_search_filter(qs, 4, "user__profile__state__name")
-        qs = self._apply_search_filter(qs, 6, "character__alliance_name")
-        qs = self._apply_search_filter(qs, 7, "character__corporation_name")
-        qs = self._apply_search_filter(
-            qs, 8, "user__profile__main_character__alliance_name"
-        )
-        qs = self._apply_search_filter(
-            qs, 9, "user__profile__main_character__corporation_name"
-        )
-        qs = self._apply_search_filter(
-            qs, 10, "user__profile__main_character__character_name"
-        )
-        qs = self._apply_search_filter(qs, 11, "unregistered")
+        # qs = self._apply_search_filter(qs, 4, "user__profile__state__name")
+        # qs = self._apply_search_filter(qs, 6, "character__alliance_name")
+        # qs = self._apply_search_filter(qs, 7, "character__corporation_name")
+        # qs = self._apply_search_filter(
+        #     qs, 8, "user__profile__main_character__alliance_name"
+        # )
+        # qs = self._apply_search_filter(
+        #     qs, 9, "user__profile__main_character__corporation_name"
+        # )
+        # qs = self._apply_search_filter(
+        #     qs, 10, "user__profile__main_character__character_name"
+        # )
+        # qs = self._apply_search_filter(qs, 11, "unregistered")
 
-        return qs
+        # return qs
 
     def _apply_search_filter(self, qs, column_num, field) -> models.QuerySet:
         my_filter = self.request.GET.get(f"columns[{column_num}][search][value]", None)
@@ -529,7 +531,7 @@ class MoonListJson(PermissionRequiredMixin, LoginRequiredMixin, BaseDatatableVie
             return qs.filter(**kwargs)
         return qs
 
-    def render_column(self, row, column) -> str:
+    def render_column(self, row, column) -> Union[str, dict]:
         if column == "id":
             return row.pk
         if column == "moon_name":
@@ -593,7 +595,7 @@ class MoonListJson(PermissionRequiredMixin, LoginRequiredMixin, BaseDatatableVie
             return details_html
         return ""
 
-    def _render_refinery(self, row, column):
+    def _render_refinery(self, row, column) -> Union[str, dict]:
         if row.has_refinery:
             refinery = row.refinery
             refinery_html = refinery.name_html()
@@ -610,7 +612,7 @@ class MoonListJson(PermissionRequiredMixin, LoginRequiredMixin, BaseDatatableVie
             return alliance_name
         if column == "refinery":
             return {"display": refinery_html, "sort": refinery_name}
-        return None
+        return ""
 
 
 @login_required
@@ -680,14 +682,14 @@ def moon_details(request, moon_pk: int):
         "total_volume_per_month": MOONMINING_VOLUME_PER_MONTH / 1000000,
     }
     if request.GET.get("new_page"):
-        context["title"] = "Moon"
+        context["title"] = _("Moon")
         context["content_file"] = "moonmining/partials/moon_details.html"
         return render(request, "moonmining/_generic_modal_page.html", context)
     return render(request, "moonmining/modals/moon_details.html", context)
 
 
 @permission_required(["moonmining.add_refinery_owner", "moonmining.basic_access"])
-@token_required(scopes=Owner.esi_scopes())
+@token_required(scopes=Owner.esi_scopes())  # type: ignore
 @login_required
 def add_owner(request, token):
     try:
@@ -706,16 +708,18 @@ def add_owner(request, token):
         )
         corporation.save()
 
-    owner, _ = Owner.objects.update_or_create(
+    owner = Owner.objects.update_or_create(
         corporation=corporation,
         defaults={"character_ownership": character_ownership},
-    )
+    )[0]
     tasks.update_owner.delay(owner.pk)
-    messages_plus.success(request, f"Update of refineres started for {owner}.")
+    messages_plus.success(request, f"Update of refineries started for {owner}.")
     if MOONMINING_ADMIN_NOTIFICATIONS_ENABLED:
         notify_admins(
-            message=("%(corporation)s was added as new owner by %(user)s.")
-            % {"corporation": owner, "user": request.user},
+            message=_(
+                "%(corporation)s was added as new owner by %(user)s."
+                % {"corporation": owner, "user": request.user}
+            ),
             title=f"{__title__}: Owner added: {owner}",
         )
     return redirect("moonmining:index")
@@ -724,7 +728,7 @@ def add_owner(request, token):
 @permission_required(["moonmining.basic_access", "moonmining.upload_moon_scan"])
 @login_required()
 def upload_survey(request):
-    context = {"page_title": "Upload Moon Surveys"}
+    context = {"page_title": _("Upload Moon Surveys")}
     if request.method == "POST":
         form = MoonScanForm(request.POST)
         if form.is_valid():
@@ -732,15 +736,15 @@ def upload_survey(request):
             tasks.process_survey_input.delay(scans, request.user.pk)
             messages_plus.success(
                 request,
-                (
-                    "Your scan has been submitted for processing. You will"
-                    "receive a notification once processing is complete."
+                _(
+                    "Your scan has been submitted for processing. "
+                    "You will receive a notification once processing is complete."
                 ),
             )
         else:
             messages_plus.error(
                 request,
-                (
+                _(
                     "Oh No! Something went wrong with your moon scan submission. "
                     "Please try again."
                 ),
@@ -777,7 +781,7 @@ def reports(request):
         except KeyError:
             ledger_last_updated = None
     context = {
-        "page_title": "Reports",
+        "page_title": _("Reports"),
         "use_reprocess_pricing": MOONMINING_USE_REPROCESS_PRICING,
         "reprocessing_yield": MOONMINING_REPROCESSING_YIELD * 100,
         "total_volume_per_month": MOONMINING_VOLUME_PER_MONTH / 1000000,
@@ -802,9 +806,11 @@ def report_owned_value_data(request):
         "refinery__owner__corporation",
         "refinery__owner__corporation__alliance",
     ).filter(refinery__isnull=False)
-    corporation_moons = defaultdict(lambda: {"moons": list(), "total": 0})
+    corporation_moons = {}
     for moon in moon_query.order_by("eve_moon__name"):
         corporation_name = moon.refinery.owner.name
+        if corporation_name not in corporation_moons:
+            corporation_moons[corporation_name] = {"moons": list(), "total": 0}
         corporation_moons[corporation_name]["moons"].append(moon)
         corporation_moons[corporation_name]["total"] += default_if_none(moon.value, 0)
 
@@ -847,7 +853,7 @@ def report_owned_value_data(request):
         data.append(
             {
                 "corporation": corporation,
-                "moon": {"display": "TOTAL", "sort": counter},
+                "moon": {"display": _("Total"), "sort": counter},
                 "region": None,
                 "rarity_class": None,
                 "value": None,